requirements.txt
asgiref==3.4.1
astroid==2.7.2
black==21.8b0
click==8.0.1
Django==3.2.6
djangorestframework==3.12.4
isort==5.9.3
lazy-object-proxy==1.6.0
mccabe==0.6.1
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.2.0
pylint==2.10.2
pylint-django==2.4.4
pylint-plugin-utils==0.6
pytz==2021.1
regex==2021.8.28
sqlparse==0.4.1
toml==0.10.2
tomli==1.2.1
typing-extensions==3.10.0.1
wrapt==1.12.1
asgiref==3.4.1
astroid==2.7.2
black==21.8b0
certifi==2021.5.30
charset-normalizer==2.0.4
click==8.0.1
Django==3.2.6
djangorestframework==3.12.4
idna==3.2
isort==5.9.3
lazy-object-proxy==1.6.0
mccabe==0.6.1
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.2.0
pylint==2.10.2
pylint-django==2.4.4
pylint-plugin-utils==0.6
pytz==2021.1
regex==2021.8.28
requests==2.26.0
sqlparse==0.4.1
toml==0.10.2
tomli==1.2.1
typing-extensions==3.10.0.1
urllib3==1.26.6
wrapt==1.12.1
asgiref==3.4.1
astroid==2.7.2
black==21.8b0
certifi==2021.5.30
charset-normalizer==2.0.4
click==8.0.1
Django==3.2.6
django-environ==0.5.0
djangorestframework==3.12.4
idna==3.2
isort==5.9.3
lazy-object-proxy==1.6.0
mccabe==0.6.1
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.2.0
pylint==2.10.2
pylint-django==2.4.4
pylint-plugin-utils==0.6
pytz==2021.1
regex==2021.8.28
requests==2.26.0
sqlparse==0.4.1
toml==0.10.2
tomli==1.2.1
typing-extensions==3.10.0.1
urllib3==1.26.6
wrapt==1.12.1
asgiref==3.4.1
astroid==2.7.2
black==21.8b0
certifi==2021.5.30
charset-normalizer==2.0.4
click==8.0.1
Django==3.2.6
django-environ==0.5.0
djangorestframework==3.12.4
idna==3.2
isort==5.9.3
lazy-object-proxy==1.6.0
mccabe==0.6.1
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.2.0
psycopg2==2.9.1
pylint==2.10.2
pylint-django==2.4.4
pylint-plugin-utils==0.6
pytz==2021.1
regex==2021.8.28
requests==2.26.0
sqlparse==0.4.1
toml==0.10.2
tomli==1.2.1
typing-extensions==3.10.0.1
urllib3==1.26.6
wrapt==1.12.1
APScheduler==3.7.0
asgiref==3.4.1
astroid==2.7.2
black==21.8b0
certifi==2021.5.30
charset-normalizer==2.0.4
click==8.0.1
Django==3.2.6
django-environ==0.5.0
djangorestframework==3.12.4
idna==3.2
isort==5.9.3
lazy-object-proxy==1.6.0
mccabe==0.6.1
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.2.0
psycopg2==2.9.1
pylint==2.10.2
pylint-django==2.4.4
pylint-plugin-utils==0.6
pytz==2021.1
regex==2021.8.28
requests==2.26.0
six==1.16.0
sqlparse==0.4.1
toml==0.10.2
tomli==1.2.1
typing-extensions==3.10.0.1
tzlocal==2.1
urllib3==1.26.6
wrapt==1.12.1
